fix(storage): index_id is an int64 now

+++ influxdb/inc/com/centreon/broker/influxdb/line_protocol_query.hh
@@ -73,7 +73,7 @@ class line_protocol_query {
   void _get_member(io::data const& d, std::ostream& is);
   void _get_string(io::data const& d, std::ostream& is);
   void _get_dollar_sign(io::data const& d, std::ostream& is);
-  uint32_t _get_index_id(io::data const& d);
+  uint64_t _get_index_id(io::data const& d);
   void _get_index_id(io::data const& d, std::ostream& is);
   void _get_host(io::data const& d, std::ostream& is);
   void _get_host_id(io::data const& d, std::ostream& is);
